+++ lib/potassium/assets/seeds/common.rb
@@ -1,0 +1,8 @@
+# Put here code to generate ONLY the necessary data to run the application.
+# Write the code in a way that can be executed multiple times without duplicating the information.
+#
+# For example:
+#
+# Country.create(name: "Chile") # BAD
+# Country.find_or_create_by(name: "Chile") # GOOD
+#