+++ lib/potassium/assets/seeds/production.rb
@@ -1,0 +1,1 @@
+# Put here the necessary code to generate data available on production environment only.